refactor(editor): reorg code structure of store package (#9525)
