+++ blocksuite/framework/store/src/consts.ts
@@ -1,7 +1,3 @@
-export const COLLECTION_VERSION = 2;
-
-export const PAGE_VERSION = 2;
-
 export const SCHEMA_NOT_FOUND_MESSAGE =
   'Schema not found. The block flavour may not be registered.';
