+++ blocksuite/framework/store/src/model/block/index.ts
@@ -1,0 +1,5 @@
+export * from './block.js';
+export * from './block-model.js';
+export * from './draft.js';
+export * from './types.js';
+export * from './zod.js';
