+++ blocksuite/framework/store/src/model/blocks/index.ts
@@ -1,0 +1,2 @@
+export * from './blocks.js';
+export * from './query.js';
