+++ blocksuite/framework/store/src/schema/index.ts
@@ -1,2 +1,1 @@
-export * from './base.js';
 export { Schema } from './schema.js';
