--- blocksuite/framework/store/src/store/doc/consts.ts
@@ -1,5 +1,0 @@
-export enum BlockViewType {
-  Bypass = 'bypass',
-  Display = 'display',
-  Hidden = 'hidden',
-}
