--- blocksuite/framework/store/src/store/doc/index.ts
@@ -1,4 +1,0 @@
-export * from './block/index.js';
-export * from './consts.js';
-export * from './doc.js';
-export * from './query.js';
