--- blocksuite/framework/store/src/store/index.ts
@@ -1,2 +1,0 @@
-export * from './doc/index.js';
-export * from './workspace.js';
