+++ blocksuite/framework/store/src/test/index.ts
@@ -1,4 +1,10 @@
-export { createAutoIncrementIdGenerator } from '../utils/id-generator.js';
+import type { IdGenerator } from '../utils/id-generator.js';
+
 export * from './test-doc.js';
 export * from './test-meta.js';
 export * from './test-workspace.js';
+
+export function createAutoIncrementIdGenerator(): IdGenerator {
+  let i = 0;
+  return () => (i++).toString();
+}
