+++ blocksuite/framework/store/src/transformer/index.ts
@@ -1,6 +1,5 @@
 export * from './assets.js';
 export * from './base.js';
-export * from './draft.js';
 export * from './job.js';
 export * from './json.js';
 export * from './middleware.js';
