+++ blocksuite/framework/store/src/yjs/utils.ts
@@ -5,3 +5,7 @@ export type SubdocEvent = {
   removed: Set<YDoc>;
   added: Set<YDoc>;
 };
+
+export interface StackItem {
+  meta: Map<'selection-state', unknown>;
+}
